Update anime-subtitle-extractor.sh refinements, and bug fixes
Added santization to the script so that subtitle titles with /'s in them get processed correctly. / is replace with _.

Added multiepisode processing, processes 4 episodes at a time now, this improves speed by a lot.

Added some coloring to more easy identify issues. Made changes to the output of the analyze task to make for easier troubleshooting.